Implement From trait for most things, address clippy complaints
This allows easy conversion into trait objects, which are required for
various interfaces and structs

Also noticed that wrapping a whole Cornell Box in a BVHNode can have
a significant adverse performance impact. Could be upwards of 10-20%
in some cases. Presumably most of that is at path tracing time but
I'm not sure yet. Generally this makes sense. If the box takes up the
whole field of view and you stick that in a bounding volume, every
ray will hit that volume, so you've gained nothing from the optimization
but presumably added some indirection overhead traversing the "hierarchy".
Worth some additional study.

Addendum: This effect is much less pronounced with the wacky cornell
box, which doesn't use the empty_cornell_box helper. This suggests that
perhaps it's the nesting that's causing problems, e.g. the tree doesn't
rebalance itself if the relationship between sub-nodes is not reflected
in the initial organization of elements.

Additionally, clean up lib.rs, remove some unused includes and address
various linter things.